Correct un-narrowing when hitting backspace
`dired-narrow--restore` is called from `dired-narrow--update` when
it's updating the narrowing of the buffer.  However, if you type
some character that doesn't occur in any file name (ex. `{`), then
backspace it, then type a character that does occur in a file
name---or just hit RET for no narrowing at all---all files will be
narrowed away when exiting dired-narrow because the `:dired-narrow`
property still exists on them from when you typed `{`.

This change just has `dired-narrow--restore` remove the
`:dired-narrow` property at the same tim as it removes the `invisible`
property.