dired-narrow: call `dired-narrow-find-file' when entering directory
Before this change, when `dired-narrow-enter-directory' was called,
the `dired-narrow--internal' function did call itself to start another
instance of dired narrow in the subdirectory.  However it did *not*
visit the directory at point, which restarted dired-narrow in the
current (already filtered) directory.

With this change `dired-narrow-find-file' is called before restarting
`dired-narrow--internal'.  This will visit the directory at point and
start another dired-narrow instance.  If the current file is not a
directory, visit the file at point.