feat(list): allow modification of display commands
Rather than hardcoding the display commands of the various provided listings,
give a set of customisation variables allowing users to change it, for instance
to use xargs(1) rather than the ls option of find(1).

This makes c7ca84c unneeded : the literal flag can now directly be added or
removed in the string describing the command.