dired-subtree--reading checks if listing already indented
If --dired is one of the listing switches, the listing will already be indented by two whitespaces. If two whitespaces are detected in the beginning of the listing, no further whitespaces will be added. Otherwise, i.e. if additional whitespaces are prepended, the directories in the subdirectory will not be identified as such and further levels of subdirectories will not work.